requirements.txt
mysql-connector-python==8.0.22
pandas==1.1.4
python-dotenv==0.15.0
SQLAlchemy==1.3.20
xlrd==1.2.0
numpy==1.19.3
